commit for 1.1.6-rc.0
