+++ .changeset/pre.json
@@ -1,0 +1,10 @@
+{
+  "mode": "pre",
+  "tag": "rc",
+  "initialVersions": {
+    "socket-warp": "1.1.5"
+  },
+  "changesets": [
+    "ten-adults-lie"
+  ]
+}
