+++ .changeset/ten-adults-lie.md
@@ -1,0 +1,5 @@
+---
+"socket-warp": patch
+---
+
+fix connector
