+++ Dockerfile
@@ -5,5 +5,6 @@ RUN cargo build --release
 
 FROM rust:latest
 WORKDIR /app
+COPY ./sw_connector .
 COPY --from=builder /build/target/release/sw_listener .
 CMD ["./sw_listener"]
