+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "socket-warp",
-  "version": "1.1.5",
+  "version": "1.1.6-rc.0",
   "private": true,
   "devDependencies": {
     "@changesets/cli": "^2.27.1",
